requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt setup.py
#
aniso8601==3.0.0          # via graphene
beautifulsoup4==4.6.0     # via bs4
bs4==0.0.1                # via gtts
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via gtts
crate==0.22.0
future==0.16.0            # via python-twitter
gevent==1.2.2
graphene==2.1
graphql-core==2.0         # via graphene, graphql-relay, graphql-server-core, webob-graphql
graphql-relay==0.4.5      # via graphene
graphql-server-core==1.0.dev20170322001  # via webob-graphql
greenlet==0.4.13          # via gevent
gtts-token==1.1.1         # via gtts
gtts==2.0.0
hupper==1.2               # via pyramid
idna==2.6                 # via requests
jinja2==2.10
markupsafe==1.0           # via jinja2
oauthlib==2.0.7           # via requests-oauthlib
pastedeploy==1.5.2        # via plaster-pastedeploy, pyramid
plaster-pastedeploy==0.5  # via pyramid
plaster==1.0              # via plaster-pastedeploy, pyramid
promise==2.1              # via graphene, graphql-core, graphql-relay, graphql-server-core
pyramid-tm==2.2
pyramid==1.9.2
python-twitter==3.4.1
repoze.lru==0.7           # via pyramid
requests-oauthlib==0.8.0  # via python-twitter
requests==2.18.4          # via gtts, gtts-token, python-twitter, requests-oauthlib
rx==1.6.1                 # via graphql-core
six==1.11.0               # via graphene, graphql-core, graphql-relay, gtts, promise
sqlalchemy==1.2.7
transaction==2.2.1        # via pyramid-tm, zope.sqlalchemy
translationstring==1.3    # via pyramid
typing==3.6.4             # via promise
urllib3==1.22             # via crate, requests
venusian==1.1.0           # via pyramid
webob-graphql==1.0
webob==1.8.1              # via pyramid, webob-graphql
zope.deprecation==4.3.0   # via pyramid
zope.interface==4.5.0     # via pyramid, transaction, zope.sqlalchemy
zope.sqlalchemy==1.0
